Rename "sthgetter" package to "sth" and shorten func name
Removes some stutter from func calls.
